refactor: make monorepo
