+++ README.md
@@ -1,1 +1,3 @@
-# nodejs-aws-shop-backend
+# AWS Cloud Developer Course Backend
+
+Backend tasks for the ["AWS Cloud Developer" course by The Rolling Scopes School](https://github.com/rolling-scopes-school/aws/tree/main/aws-developer)
