+++ product_service/.prettierrc
