+++ product_service/src/cdk/lib/handlers/dynamo/get_products_by_id.ts
@@ -15,7 +15,7 @@ import {
   AppResponse,
   ErrorData,
   buildServerErrorResponse,
-} from '../../../../../utils/utils';
+} from '../../../../utils/utils';
 
 interface StocksGetItemCommandOutput extends GetItemCommandOutput {
   Item?: StockItemDynamoDb;
