+++ product_service/src/cdk/lib/handlers/rds/get_products_by_id.ts
@@ -4,7 +4,7 @@ import {
   AppResponse,
   ErrorData,
   buildServerErrorResponse,
-} from '../../../../../utils/utils';
+} from '../../../../utils/utils';
 import { LambdaEventDetail } from '../types';
 import { getPgClient } from './utils';
