+++ product_service/src/cdk/lib/handlers/rds/utils.ts
