+++ product_service/src/cdk/lib/handlers/types.ts
