+++ product_service/src/cdk/lib/handlers/validation.ts
