+++ product_service/src/cdk/lib/init_dynamo_api.ts
