+++ product_service/src/cdk/lib/mocks/products.ts
