+++ product_service/src/cdk/lib/mocks/stocks.ts
