+++ product_service/src/cdk/lib/models/product.ts
