+++ product_service/src/cdk/lib/models/product_item_dynamodb.ts
