+++ product_service/src/cdk/lib/models/product_service_props.ts
