+++ product_service/src/cdk/lib/models/product_with_stock.ts
