+++ product_service/src/cdk/lib/models/stock.ts
