+++ product_service/src/cdk/lib/models/stock_item_dynamodb.ts
