+++ product_service/src/types/index.d.ts
