+++ product_service/src/utils/load_env.ts
