+++ product_service/src/utils/rds_utils.ts
