+++ product_service/src/utils/utils.ts
