--- src/utils/constants.ts
@@ -1,2 +1,0 @@
-export const PRODUCTS_TABLE_NAME = 'products';
-export const STOCKS_TABLE_NAME = 'stocks';
